requirements.txt
easyocr
Pillow
matplotlib
tk